postprocess-lrs-cloud-function/pyproject.toml
[tool.poetry]
name = "postprocess-lrs-cloud-function"
version = "0.1.0"
description = ""
authors = ["Marcus Vinicius Monteiro de Souza <marcusmonteirodesouza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
functions-framework = "^3.5.0"
google-cloud-storage = "^2.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
